module level logging control

+++ scripts/run_inference.py
@@ -32,6 +32,7 @@ logging.basicConfig(
     datefmt="%Y-%m-%d %H:%M:%S")
 
 logger = logging.getLogger(__name__)
+logging.getLogger('scripts.process.process_helpers').setLevel(logging.WARNING)
 
 device = pick_device()
 
@@ -50,10 +51,10 @@ def create_weight_matrix(tile_size, overlap_size):
 
 
 def make_prediction_tiles(tile_folder, metadata, model, device, threshold, ):
-    print(f'---ORIGINAL PREDICTIONS FUNCTION')
+    logger.info(f'---ORIGINAL PREDICTIONS FUNCTION')
     predictions_folder = Path(tile_folder).parent / f'{tile_folder.stem}_predictions'
     if predictions_folder.exists():
-        print(f"--- Deleting existing predictions folder: {predictions_folder}")
+        logger.info(f"--- Deleting existing predictions folder: {predictions_folder}")
         # delete the folder and create a new one
         shutil.rmtree(predictions_folder)
     predictions_folder.mkdir(exist_ok=True)
@@ -88,7 +89,7 @@ def make_prediction_tiles(tile_folder, metadata, model, device, threshold, ):
 def make_prediction_tiles_new(tile_folder, metadata, model, device, threshold, stride):
     predictions_folder = Path(tile_folder).parent / f'{tile_folder.stem}_predictions'
     if predictions_folder.exists():
-        print(f"--- Deleting existing predictions folder: {predictions_folder}")
+        logger.info(f"--- Deleting existing predictions folder: {predictions_folder}")
         shutil.rmtree(predictions_folder)
     predictions_folder.mkdir(exist_ok=True)
 
@@ -160,15 +161,15 @@ def stitch_tiles(metadata, prediction_tiles, save_path, image):
         transform = src.transform
         crs = src.crs
         height, width = src.shape
-        print('>>>src shape:',src.shape)
+        logger.info(f'src shape: {src.shape}')
     
         # INITIALIZE THE STITCHED IMAGE AND COUNT
         # give stitched_image the same crs, transform and shape as the source image
         stitched_image = np.zeros((height, width))
-        # print(">>>stitched_image dtype:", stitched_image.dtype)
-        print(">>>stitched_image shape:", stitched_image.shape)
-        #print unique values in the stitched image
-        # print(f'>>>unique values in empty stitched image: {np.unique(stitched_image)}')
+        # logger.info("stitched_image dtype:", stitched_image.dtype)
+        logger.info(f"stitched_image shape: { stitched_image.shape}")
+        #logger.info unique values in the stitched image
+        # logger.info(f'unique values in empty stitched image: {np.unique(stitched_image)}')
 
     for tile_info in tqdm(metadata, desc="Stitching tiles"):
         tile_name = tile_info["tile_name"]
@@ -182,9 +183,9 @@ def stitch_tiles(metadata, prediction_tiles, save_path, image):
         # Load the tile
         with rasterio.open(tile) as src:
             tile = src.read(1).astype(np.float32)
-            # Debugging: Print tile info and shapes
-            # print(f">>>Tile shape: {tile.shape}")
-        # print(f">>> Tile info: {tile_info}")
+            # Debugging: logger.info tile info and shapes
+            # logger.info(f"Tile shape: {tile.shape}")
+        # logger.info(f" Tile info: {tile_info}")
 
         # Extract the relevant slice from the stitched image
         stitched_slice = stitched_image[y_start:y_end, x_start:x_end]
@@ -195,18 +196,18 @@ def stitch_tiles(metadata, prediction_tiles, save_path, image):
         if stitched_slice.shape != tile.shape:
             if (stitched_slice.shape[0] == 0) or (stitched_slice.shape[1] == 0):
                 continue
-            print(f"---Mismatch: Stitched slice shape: {stitched_slice.shape}, ---Tile shape: {tile.shape}")
+            logger.info(f"---Mismatch: Stitched slice shape: {stitched_slice.shape}, ---Tile shape: {tile.shape}")
             slice_height, slice_width = stitched_slice.shape
             tile = tile[:slice_height, :slice_width]  # Crop tile to match slice
-            # Debugging: Print the new tile shape
-            print(f">>>New tile shape: {tile.shape}")
+            # Debugging: logger.info the new tile shape
+            logger.info(f"New tile shape: {tile.shape}")
 
 
         # Add the tile to the corresponding position in the stitched image
         stitched_image[y_start:y_end, x_start:x_end] += tile
-        # PRINT STITCHED IMAGE SIZE
-        # print(f">>>Stitched image shape: {stitched_image.shape}")
-    print(f'---crs: {crs}')
+        # logger.info STITCHED IMAGE SIZE
+        # logger.info(f"Stitched image shape: {stitched_image.shape}")
+    logger.info(f'---crs: {crs}')
     # Save the stitched image as tif, as save_path
     with rasterio.open(
         save_path,
@@ -221,7 +222,7 @@ def stitch_tiles(metadata, prediction_tiles, save_path, image):
     ) as dst:
         dst.write(stitched_image, 1)
     # with rasterio.open(save_path) as src:
-    #     print("No-data value:", src.nodata)
+    #     logger.info("No-data value:", src.nodata)
         
     return stitched_image
 
@@ -249,20 +250,20 @@ def main(test=False):
     # plt.imshow(wm, cmap="viridis")
     # plt.colorbar()
     # plt.show()
-    print(f'>>>test mode = {test}')
+    logger.info(f'test mode = {test}')
 
 
     # VARIABLES................................................................
     norm_func = 'logclipmm_g' # 'mm' or 'logclipmm'
     stats = None
-    MAKE_TIFS = True
-    MAKE_DATAARRAY= True
+    MAKE_TIFS = 1
+    MAKE_DATAARRAY= 1
     # stride = tile_size
     ############################################################################
     # DEFINE PATHS
     # DEFINE THE WORKING FOLDER FOR I/O
     predict_input = Path("/Users/alexwebb/laptop_coding/floodai/UNOSAT_FloodAI_v2/data/4final/predict_input")
-    print(f'>>>working folder: {predict_input}')
+    logger.info(f'working folder: {predict_input.name}')
     if path_not_exists(predict_input):
         return
     
@@ -279,12 +280,12 @@ def main(test=False):
         # Normalize all paths in the config
         image = check_single_input_filetype(predict_input, 'image', '.tif')
         if image is None:
-            print(f"---No input image found in {predict_input}")
+            logger.info(f"---No input image found in {predict_input}")
             return
         else:
-            print(f'>>>found input image: {image.name}')
+            logger.info(f'found input image: {image.name}')
         output_folder = predict_input
-        output_filename = '_x'
+        output_filename = '_name'
         # analysis_extent = Path('Users/alexwebb/floodai/UNOSAT_FloodAI_v2/data/4final/predict_INPUT/extent_INPUT')  
 
     # READ CONFIG
@@ -302,20 +303,19 @@ def main(test=False):
 
     stride = tile_size # STRIDE FOR TILING, SAME AS TILE SIZE
 
-    # print(f'>>> config = {config}')
-    print(f'>>>image: {image}')
-    print(f'>>>threshold: {threshold}') 
-    print(f'>>>tile_size: {tile_size}')
-    print(f'>>>output_folder= {output_folder}')
-    print(f'>>>output_filename= {output_filename}')
-    # print(f'>>>alalysis_extent= {analysis_extent}')
-    # print(f'>>> IF TRAINING: CHECK LAYERDICT NAMES=FILENAMES IN FOLDER <<<')
+    # logger.info(f' config = {config}')
+    logger.info(f'threshold: {threshold}') 
+    logger.info(f'tile_size: {tile_size}')
+    logger.info(f'output_folder= {output_folder.name}')
+    logger.info(f'output_filename= {output_filename}')
+    # logger.info(f'alalysis_extent= {analysis_extent}')
+    # logger.info(f' IF TRAINING: CHECK LAYERDICT NAMES=FILENAMES IN FOLDER <<<')
     # FIND THE CKPT
     ckpt = next(ckpt_path.rglob("*.ckpt"), None)
     if ckpt is None:
-        print(f"---No checkpoint found in {ckpt_path}")
+        logger.info(f"---No checkpoint found in {ckpt_path}")
         return
-    print(f'>>>ckpt: {ckpt.name}')
+    logger.info(f'ckpt: {ckpt.name}')
 
     # poly = check_single_input_filetype(predict_input,  'poly', '.kml')
     # if poly is None:
@@ -324,54 +324,54 @@ def main(test=False):
     # GET REGION CODE FROM MASK TODO
     # sensor = image.parents[1].name.split('_')[:1]
     sensor = 'sensor'
-    # print(f'>>>datatype= ',sensor[0])
+    # logger.info(f'datatype= ',sensor[0])
     # date = image.parents[1].name.split('_')[0]
     date = 'date'
-    # print(f'>>>date= ',date)
+    # logger.info(f'date= ',date)
     # image_code = "_".join(image.parents[3].name.split('_')[4:])
     # image_code = "_".join(image.parents[1].name.split('_')[1])
     parts = image.name.split('_')
     image_code = "_".join(parts[:-1])
-    # print(f'>>>image_code= ',image_code)
+    # logger.info(f'image_code= ',image_code)
     save_path = output_folder / f'{sensor}_{image_code}_{date}_{tile_size}_{threshold}{output_filename}WATER_AI.tif'
 
-    print(f'>>>save_path: {save_path.name}')
+    logger.info(f'output filename: {save_path.name}')
     if save_path.exists():
-        print(f"---overwriting existing file! : {save_path}")
+        logger.info(f"---overwriting existing file! : {save_path}")
         # try:
-        #     print(f"--- Deleting existing prediction file: {save_path}")
+        #     logger.info(f"--- Deleting existing prediction file: {save_path}")
         #     save_path.unlink()
         # except Exception as e:
-        #     print(f"--- Error deleting existing prediction file: {e}")
+        #     logger.info(f"--- Error deleting existing prediction file: {e}")
         #     return
 
     # CREATE THE EXTRACTED FOLDER
     extracted = predict_input / f'{image_code}_extracted'
 
     
-    print(f'>>> MAKE_TIFS = {MAKE_TIFS}')
+    logger.info(f' MAKE_TIFS = {MAKE_TIFS}')
 
     if MAKE_TIFS:
         if extracted.exists():
-            # print(f"--- Deleting existing extracted folder: {extracted}")
+            # logger.info(f"--- Deleting existing extracted folder: {extracted}")
             # delete the folder and create a new one
             shutil.rmtree(extracted)
         extracted.mkdir(exist_ok=True)
 
         # CHANGE DATATYPE TO FLOAT32
-        print('>>>CHANGING DATATYPE')
+        logger.info('CHANGING DATATYPE')
         image_32 = extracted / f'{image_code}_32.tif'
         make_float32_inf(image, image_32)
-        # print_tiff_info_TSX(image_32, 1)
+        # logger.info_tiff_info_TSX(image_32, 1)
 
         # RESAMPLE TO 2.5
-        # print('>>>RESAMPLING')
+        # logger.info('RESAMPLING')
         # resamp_image = extracted / f'{image_32.stem}_resamp'
         # resample_tiff_gdal(image_32, resamp_image, target_res=2.5)
-        # print_tiff_info_TSX(resamp_image, 2)
+        # logger.info_tiff_info_TSX(resamp_image, 2)
 
         # with rasterio.open(image) as src:
-            # print(f'>>>src shape= ',src.shape)
+            # logger.info(f'src shape= ',src.shape)
 
         # SORT OUT ANALYSIS EXTENT
 
@@ -380,10 +380,10 @@ def main(test=False):
         # rasterize_kml_rasterio( poly, ex_extent, pixel_size=0.0001, burn_value=1)
 
         # REPROJECT IMAGE
-        print('>>>REPROJECTING')
+        logger.info('REPROJECTING')
         final_image = extracted / 'final_image.tif'
         reproject_to_4326_gdal(image_32, final_image, resampleAlg = 'bilinear')
-        # print_tiff_info_TSX(final_image, 3)
+        # logger.info_tiff_info_TSX(final_image, 3)
 
         # reproj_extent = extracted / f'{image_code}_4326_extent.tif'
         # reproject_to_4326_gdal(ex_extent, reproj_extent)
@@ -399,11 +399,14 @@ def main(test=False):
     if MAKE_DATAARRAY:
         create_event_datacube_TSX_inf(predict_input, image_code)
 
-    cube = next(predict_input.rglob("*.nc"), None)  
+    cube = next(predict_input.rglob("*.nc"), None)
+    if cube is None:
+        logger.info(f"---No data cube found in {predict_input.name}")
+        return  
     save_tiles_path = predict_input /  f'{image_code}_tiles'
 
     if save_tiles_path.exists():
-        # print(f">>> Deleting existing tiles folder: {save_tiles_path}")
+        # logger.info(f" Deleting existing tiles folder: {save_tiles_path}")
         # delete the folder and create a new one
         shutil.rmtree(save_tiles_path)
         save_tiles_path.mkdir(exist_ok=True, parents=True)
@@ -411,8 +414,8 @@ def main(test=False):
 
     # DO THE TILING
     tiles, metadata = tile_datacube_rxr_inf(cube, save_tiles_path, tile_size=tile_size, stride=stride, norm_func=norm_func, stats=stats, percent_non_flood=0, inference=True) 
-    # print(f">>>{len(tiles)} tiles saved to {save_tiles_path}")
-    # print(f">>>{len(metadata)} metadata saved to {save_tiles_path}")
+    # logger.info(f"{len(tiles)} tiles saved to {save_tiles_path}")
+    # logger.info(f"{len(metadata)} metadata saved to {save_tiles_path}")
     # metadata = Path(save_tiles_path) / 'tile_metadata.json'
 
     # INITIALIZE THE MODEL
@@ -440,8 +443,8 @@ def main(test=False):
 
     # STITCH PREDICTION TILES
     prediction_img = stitch_tiles(metadata, prediction_tiles, save_path, final_image)
-    # print prediction_img size
-    # print(f'>>>prediction_img shape:',prediction_img.shape)
+    # logger.info prediction_img size
+    # logger.info(f'prediction_img shape:',prediction_img.shape)
     # display the prediction mask
     # plt.imshow(prediction_img, cmap='gray')
     # plt.show()
@@ -453,7 +456,7 @@ def main(test=False):
 
     end = time.time()
     # time taken in minutes to 2 decimal places
-    print(f"Time taken: {((end - start) / 60):.2f} minutes")
+    logger.info(f"Time taken: {((end - start) / 60):.2f} minutes")
 
 if __name__ == "__main__":
     main()
